+++ tests/ext/test_ase_pyg.py
@@ -24,7 +24,7 @@ def test_PESCalculator_and_M3GNetCalculator(MoS, caplog):
     adaptor = AseAtomsAdaptor()
     s_ase = adaptor.get_atoms(MoS)  # type: ignore
 
-    ff = matgl.load_model("Materialyze/TensorNet-PES-MatPES-PBE-2025.2")
+    ff = matgl.load_model("materialyze/TensorNet-PES-MatPES-PBE-2025.2")
     ff.calc_hessian = True
 
     # ------------------------------------------------------------------
@@ -104,7 +104,7 @@ def test_PESCalculator_and_M3GNetCalculator(MoS, caplog):
 def test_PESCalculator_mol(AcAla3NHMe):
     adaptor = AseAtomsAdaptor()
     mol = adaptor.get_atoms(AcAla3NHMe)
-    ff = matgl.load_model("Materialyze/TensorNet-PES-MatPES-PBE-2025.2")
+    ff = matgl.load_model("materialyze/TensorNet-PES-MatPES-PBE-2025.2")
     calc = PESCalculator(potential=ff)
     mol.set_calculator(calc)
     assert isinstance(mol.get_potential_energy(), float)
@@ -113,7 +113,7 @@ def test_PESCalculator_mol(AcAla3NHMe):
 
 
 def test_Relaxer(MoS):
-    pot = matgl.load_model("Materialyze/TensorNet-PES-MatPES-PBE-2025.2")
+    pot = matgl.load_model("materialyze/TensorNet-PES-MatPES-PBE-2025.2")
     r = Relaxer(pot)
     results = r.relax(MoS, traj_file="MoS_relax.traj")
     s = results["final_structure"]
@@ -160,7 +160,7 @@ def test_get_graph_from_atoms_mol():
 
 
 def test_molecular_dynamics(MoS2):
-    pot = matgl.load_model("Materialyze/TensorNet-PES-MatPES-PBE-2025.2")
+    pot = matgl.load_model("materialyze/TensorNet-PES-MatPES-PBE-2025.2")
     for ensemble in [
         "nvt",
         "nve",