+++ tests/test_integration.py
@@ -11,15 +11,10 @@ import matgl
 
 @pytest.mark.skipif(matgl.config.BACKEND != "DGL", reason="Only works with DGL.")
 def test_form_e(LiFePO4):
-    model = matgl.load_model("M3GNet-MP-2018.6.1-Eform")
+    model = matgl.load_model("M3GNet-Eform-MP-2018.6.1")
     for _i in range(3):
         # This loop ensures there is no stochasticity in the prediction, a problem in v1 of the models.
         assert model.predict_structure(LiFePO4) == pytest.approx(-2.5489, 3)
-
-
-@pytest.mark.skipif(matgl.config.BACKEND != "DGL", reason="Only works with DGL.")
-def test_chgnet_loading():
-    matgl.load_model("CHGNet-MatPES-PBE-2025.2.10-2.7M-PES")
 
 
 @pytest.mark.skipif(os.getenv("CI") == "true" or matgl.config.BACKEND != "DGL", reason="Unreliable in CI environments.")